+++ docs/sapnwrfc.stubs.php
@@ -4,12 +4,12 @@
  * This is a stub file of the extensions public interface to enable
  * code completion in IDEs.
  */
-
 namespace SAPNWRFC;
 
-class Exception extends \RuntimeException {
+class Exception extends \RuntimeException
+{
     /**
-     * Detailed RFC/ABAP error information
+     * Detailed RFC/ABAP error information.
      *
      * $errorInfo contains at least the "code", "key" and "message" keys from
      * the RFC library.
@@ -18,7 +18,7 @@ class Exception extends \RuntimeException {
      *    "abapMsgClass", "abapMsgType", "abapMsgNumber", "abapMsgV1",
      *    "abapMsgV2", "abapMsgV3", "abapMsgV4"
      *
-     * @var array $errorInfo
+     * @var array
      */
     public $errorInfo;
 
@@ -27,14 +27,21 @@ class Exception extends \RuntimeException {
      *
      * @returns array|null
      */
-    public function getErrorInfo() {}
+    public function getErrorInfo()
+    {
+    }
 }
 
-class ConnectionException extends Exception {}
+class ConnectionException extends Exception
+{
+}
 
-class FunctionCallException extends Exception {}
+class FunctionCallException extends Exception
+{
+}
 
-class Connection {
+class Connection
+{
     const TRACE_LEVEL_OFF = 0;
     const TRACE_LEVEL_BRIEF = 1;
     const TRACE_LEVEL_VERBOSE = 2;
@@ -44,43 +51,59 @@ class Connection {
      * Connect to the system using the given parameters.
      *
      * @param array $parameters Connection parameters.
+     *
      * @throws ConnectionException if the connection fails.
      */
-    public function __construct(array $parameters) {}
+    public function __construct(array $parameters)
+    {
+    }
 
     /**
      * Get the connection attributes.
      *
      * @return array Array of connection attributes.
+     *
      * @throws ConnectionException if the connection attributes could not be
      *                             fetched.
      */
-    public function attributes() {}
+    public function attributes()
+    {
+    }
 
     /**
      * @return bool True if ping successful.
+     *
      * @throws ConnectionException if the ping failed.
      */
-    public function ping() {}
+    public function ping()
+    {
+    }
 
     /**
      * Lookup a RFC function and return a RemoteFunction object.
      *
      * @param string $functionnName Name of the function.
+     *
      * @return RemoteFunction A RemoteFunction class for the RFC function.
+     *
      * @throws FunctionCallException if the lookup fails or an error is
      *                               returned during parameter parsing.
      */
-    public function getFunction($functionName) {}
+    public function getFunction($functionName)
+    {
+    }
 
     /**
      * Close the connection.
      *
      * @return bool True if the connection was closed, false if the connection
      *              is closed already.
+     *
      * @throws ConnectionException if the connection could not be closed.
      */
-    public function close() {}
+    public function close()
+    {
+    }
 
     /**
      * Sets the path to the sapnwrfc.ini file.
@@ -88,10 +111,14 @@ class Connection {
      * By default, the INI file is searched for in the current directory.
      *
      * @param string $path Path to the sapnwrfc.ini file.
+     *
      * @return bool True if path was set.
+     *
      * @throws ConnectionException if path could not be set.
      */
-    public static function setIniPath($path) {}
+    public static function setIniPath($path)
+    {
+    }
 
     /**
      * Reload the INI file.
@@ -100,50 +127,92 @@ class Connection {
      * Connection::setIniFile() or in the current directory.
      *
      * @return bool True if INI file was reloaded.
+     *
      * @throws ConnectionException if the INI file could not be reloaded.
      */
-    public static function reloadIniFile() {}
+    public static function reloadIniFile()
+    {
+    }
+
+    /**
+     * Set trace directory.
+     *
+     * @param string $path Path to trace directory (must exist).
+     *
+     * @return bool True if path was set.
+     *
+     * @throws ConnectionException if path could not be set.
+     */
+    public static function setTraceDir($path)
+    {
+    }
+
+    /**
+     * Set trace level.
+     *
+     * @param string $level Trace level.
+     *
+     * @return bool True if level was set.
+     *
+     * @throws ConnectionException if level could not be set.
+     */
+    public static function setTraceLevel($level)
+    {
+    }
 
     /**
      * Get the extension version.
      *
      * @return string The extension version.
      */
-    public static function version() {}
+    public static function version()
+    {
+    }
 
     /**
      * Get the RFC SDK version.
      *
      * @return string The RFC SDK version.
      */
-    public static function rfcVersion() {}
+    public static function rfcVersion()
+    {
+    }
 }
 
-class RemoteFunction {
+class RemoteFunction
+{
     /**
      * Invoke the RFC function.
      *
      * @param array $parameters Function parameters.
      * @returns array Return value from the backend.
+     *
      * @throws FunctionCallException if any error occurs during execution.
      */
-    public function invoke(array $parameters) {}
+    public function invoke(array $parameters)
+    {
+    }
 
     /**
      * Make a parameter active or inactive.
      *
      * @param string $parameterName The parameter to modify.
-     * @param bool $isActive True to activate the parameter, false to deactivate.
-     * @return void
+     * @param bool   $isActive      True to activate the parameter, false to deactivate.
+     *
      * @throws FunctionCallException if the parameter status could not be set.
      */
-    public function setParameterActive($parameterName, $isActive) {}
+    public function setParameterActive($parameterName, $isActive)
+    {
+    }
 
     /**
      * Check if a parameter is active or inactive.
      *
      * @param string $parameterName The parameter to check.
+     *
      * @return bool True if parameter is active, false if not.
      */
-    public function isParameterActive($parameterName) {}
+    public function isParameterActive($parameterName)
+    {
+    }
 }