+++ CHANGELOG.md
@@ -8,7 +8,9 @@ All notable changes to this project will be documented in this file.
 - A lot more tests
 - `getErrorInfo()` for exception classes (#14)
 - Argument info and types to enable reflection and type safety
-- Improve parameter handling
+- Improve parameter handling (#15)
+- `setTraceDir()` and `setTraceLevel()` for setting the trace directory and level (#17)
+-  for setting the trace level
 
 ### Changed
 - `TRACE_LEVEL_*` constants are now integers